+++ comm_backend/CMakeLists.txt
@@ -4,11 +4,14 @@ include_directories(${GLOO_INCLUDE_DIRS})
 file(GLOB HEADLM_COMM_BACKEND_SRCS
   c10dOps.cpp
   HeadLmProcessGroup.cpp
+  adapter/CpuBackend.cpp
   )
 add_library(headlm_comm_backend SHARED ${HEADLM_COMM_BACKEND_SRCS})
 message(STATUS "Torch TORCH_INCLUDE_DIRS: ${TORCH_INCLUDE_DIRS}")
 target_include_directories(headlm_comm_backend PUBLIC ${TORCH_INCLUDE_DIRS})
 target_link_libraries(headlm_comm_backend PUBLIC Python3::Python torch)
+target_link_libraries(headlm_comm_backend PUBLIC cpu_backend_adapter)
+
 
 file(GLOB HEADLM_COMM_BACKEND_TEST_SRCS HeadLmProcessGroupTest.cpp)
 add_executable(headlm_comm_backend_test ${HEADLM_COMM_BACKEND_TEST_SRCS})
@@ -23,4 +26,4 @@ add_executable(headlm_comm_backend_test ${HEADLM_COMM_BACKEND_TEST_SRCS})
 target_link_libraries(headlm_comm_backend_test headlm_comm_backend Python3::Python torch)
 add_test(headlm_comm_backend_test headlm_comm_backend_test)
 
-add_subdirectory(comm_backend)
+add_subdirectory(adapter)
