+++ comm_backend/adapter/README.txt
@@ -1,0 +1,3 @@
+An `adapter` in HeadLm is a Pytorch ProcessGroup backend which runs on specific hardware device.
+
+The HeadLmProcessGroup will return the adapter based on the hardware it runs on.
