+++ setup.py
@@ -13,10 +13,14 @@ import torch.utils
 torch_install_path = os.path.dirname(os.path.dirname(torch.utils.cmake_prefix_path))
 torch_lib_path = os.path.join(torch_install_path, 'lib')
 
-sources = ["comm_backend/HeadLmProcessGroup.cpp"]
+sources = [
+    "comm_backend/HeadLmProcessGroup.cpp",
+    "comm_backend/adapter/CpuBackend.cpp"
+    ]
 library_dirs = [torch_lib_path] 
 include_dirs = [
     f"{os.path.dirname(os.path.abspath(__file__))}/comm_backend/",
+    f"{os.path.dirname(os.path.abspath(__file__))}/comm_backend/adapter",
     f"{torch_install_path}/include/"
 ]
 
@@ -25,14 +29,14 @@ if torch.cuda.is_available():
         name="headlm_comm",
         sources=sources,
         include_dirs=include_dirs,
-        extra_compile_args=['-DUSE_C10D_GLOO=1', '-DUSE_GLOG']
+        extra_compile_args=['-DUSE_C10D_GLOO=1', '-DUSE_C10D_NCCL=1', '-DUSE_GLOG']
     )
 else:
     module = cpp_extension.CppExtension(
         name="headlm_comm",
         sources=sources,
         include_dirs=include_dirs,
-        extra_compile_args=['-DUSE_C10D_GLOO=1', '-DUSE_GLOG']
+        extra_compile_args=['-DUSE_C10D_GLOO=1', '-DUSE_C10D_NCCL=1', '-DUSE_GLOG']
     )
 
 '''
